Updated attribute and corresponding functions
Attributes now have a public readonly enum (replacing the numeric bool) that corresponds to whether it is binary numeric, numeric, or categorical

Updated Functions:

Constructor:
-Saw that one coming

GetVarID(String Variant);
-Uses a switch.
-Now returns a double (helpful for pure numberic attributes)
-Being able to return anything would actually be useful. Too bad this isnt't Python.

Updated Cases:

AttributeVals is now an array of doubles

Updated ID3:

Now throws exceptions when you try to use pure numeric attributes on it. Also casts attributevals to be integers when using it and assumes that the attributes are not purely numerical.

+++ General_Tools/Case.cs
@@ -13,7 +13,7 @@ namespace General_Tools
     public class Case
     {
         public readonly int ID;
-        public readonly int[] AttributeVals; //Attribute vals are important to know, and don't need to be changed after creation
+        public readonly double[] AttributeVals; //Attribute vals are important to know, and don't need to be changed after creation
         private double Weight;
         //public readonly String[] AttributeOrder;
 
@@ -21,7 +21,7 @@ namespace General_Tools
         /// Creates a new Case identified by an input id to contain input attribute values. The data is not self describing, and 
         /// requires a similarly ordered list of attributes (ordered by ID) to make sense of it.
         /// </summary>
-        public Case(int id, int[] attributevals/*, String[] order*/)
+        public Case(int id, double[] attributevals/*, String[] order*/)
         {
             ID = id;
             AttributeVals = attributevals;
@@ -33,7 +33,7 @@ namespace General_Tools
         /// Creates a new Case identified by an input id to contain input attribute values. Also records a double representing weight
         /// to act as part of the case of an orginal case lacking a value.
         /// </summary>
-        public Case(int id, int[] attributevals/*, String[] order*/, double weight)
+        public Case(int id, double[] attributevals/*, String[] order*/, double weight)
         {
             ID = id;
             AttributeVals = attributevals;
